Run MLX stage A via run_suite.py registration
